modify art page to display thumbnails
thumbnails are currently created using shell script and require
imagemagick installed on the server. i know it's stupid, but i think
that including thumbnails to repo is also stupid.

maybe later i replace this script with actual jekyll plugin, but for
this i'll need to familiarize myself with ruby.
but now i use following addition to the git hook on the server:

```
[ -f "$TMP_GIT_CLONE/art/thumbs.sh" ] && {
    dir="$PWD"
    cd "$TMP_GIT_CLONE/art/" && sh thumbs.sh && cd "$dir"
}
```

it happens after cloning to remote repo and before building with bundle.

overall, this is a bodge now, it recompiles all thumbnails every time
due to quirks in my current setup.